+++ packages/tautulli/lib/src/models/session.dart
@@ -1,5 +1,3 @@
-part of tautulli;
-
 /// Model for a single activity session in Tautulli.
 /// 
 /// Typically contained with [TautulliActivity], fetched from `get_activity`.