+++ packages/tautulli/CHANGELOG.md
@@ -1,5 +1,8 @@
+# 0.0.1-pre.3
+- Implemented [update](https://github.com/Tautulli/Tautulli/blob/master/API.md#update)
+- Implemented [update_check](https://github.com/Tautulli/Tautulli/blob/master/API.md#update_check)
 # 0.0.1-pre.2
 - Implemented HTTP client
-- Implemented `arnold()`
+- Implemented [arnold](https://github.com/Tautulli/Tautulli/blob/master/API.md#arnold)
 # 0.0.1-pre.1
 - Initial Release